Add pluggable session support, make Parser user-friendly
