+++ eg/brain/admin.rive
@@ -16,5 +16,5 @@
 
 > object shutdown python
 	# Shut down
-	exit(0)
+	quit()
 < object
