+++ rivescript/utils.py
@@ -46,7 +46,7 @@ def is_atomic(trigger):
     # Atomic triggers don't contain any wildcards or parenthesis or anything
     # of the sort. We don't need to test the full character set, just left
     # brackets will do.
-    special = ['*', '#', '_', '(', '[', '<', '@']
+    special = ['*', '#', '_', '(', '[', '<', '{', '@']
     for char in special:
         if char in trigger:
             return False
